chore(runtime): separate configs
